Library of component separation setups
We have two requirements for the user interface to compoenent
separation:
1) it has to be very flexible, to allow advanced usage
2) it has to extremely user friendly, to avoid overwelming the users that
   just look for some 0th order component separation.

Because of 1) we wanted the user to interface directly with the routines
that create A_ev, with the likelihood maximizer, with the low level
objects (e.g. the free parameters in the form of a single array).
All of this made the attempts to address 2) combersome and unsatisfactory.

This commit gives up on having both 1) and 2).
a) The Components now have predefined reference values hard-coded in the
   class definition. The user doesn't have to specify them in order to
   build x0
b) The user has a set of component separation setups at his disposal in
   separation_recipies.py. The most simple one requires just the list of
   the components, the instrument and the data vector. The routines
   hides all the complications due to A_ev, x0 etc. and just returns the
   results of component separation in a user friendly format.